requirements.txt
torch
transformers
python-docx
fpdf
PyPDF2
tk